+++ src/main/scala/dev/kamu/cli/MetadataRepository.scala
@@ -10,7 +10,13 @@ package dev.kamu.cli
 
 import java.io.PrintWriter
 
-import dev.kamu.core.manifests.{Dataset, DatasetID, Manifest, VolumeLayout}
+import dev.kamu.core.manifests.{
+  Dataset,
+  DatasetID,
+  Volume,
+  VolumeID,
+  VolumeLayout
+}
 import dev.kamu.cli.utility.DependencyGraph
 import org.apache.hadoop.fs.{FileSystem, Path}
 import dev.kamu.core.manifests.parsing.pureconfig.yaml
@@ -27,17 +33,30 @@ class MetadataRepository(
 ) {
   private val logger = LogManager.getLogger(getClass.getName)
 
+  protected val datasetRepo =
+    new GenericResourceRepository[Dataset, DatasetID](
+      fileSystem,
+      workspaceLayout.datasetsDir,
+      "dataset",
+      DatasetID,
+      (ds: Dataset) => ds.id
+    )
+
+  protected val volumeRepo =
+    new GenericResourceRepository[Volume, VolumeID](
+      fileSystem,
+      workspaceLayout.volumesDir,
+      "volume",
+      VolumeID,
+      (vol: Volume) => vol.id
+    )
+
   ////////////////////////////////////////////////////////////////////////////
   // Datasets
   ////////////////////////////////////////////////////////////////////////////
 
   def getDataset(id: DatasetID): Dataset = {
-    val path = getDatasetDefinitionPath(id)
-
-    if (!fileSystem.exists(path))
-      throw new DoesNotExistsException(id)
-
-    loadDatasetFromFile(path)
+    datasetRepo.getResource(id)
   }
 
   def getDatasetsInDependencyOrder(
@@ -57,85 +76,35 @@ class MetadataRepository(
   }
 
   def getAllDatasetIDs(): Seq[DatasetID] = {
-    fileSystem
-      .listStatus(workspaceLayout.datasetsDir)
-      .map(_.getPath.getName)
-      .map(filename => filename.substring(0, filename.length - ".yaml".length))
-      .map(DatasetID)
+    datasetRepo.getAllResourceIDs()
   }
 
   def getAllDatasets(): Seq[Dataset] = {
-    val manifestFiles = fileSystem
-      .listStatus(workspaceLayout.datasetsDir)
-      .map(_.getPath)
-
-    manifestFiles.map(loadDatasetFromFile)
-  }
-
-  def loadDatasetFromFile(p: Path): Dataset = {
-    val inputStream = fileSystem.open(p)
-    try {
-      yaml.load[Manifest[Dataset]](inputStream).content
-    } catch {
-      case e: Exception =>
-        logger.error(s"Error while loading data source manifest from file: $p")
-        throw e
-    } finally {
-      inputStream.close()
-    }
+    datasetRepo.getAllResources()
   }
 
   def loadDatasetFromURI(uri: URI): Dataset = {
-    uri.getScheme match {
-      case "https" => loadDatasetFromURL(uri.toURL)
-      case "http"  => loadDatasetFromURL(uri.toURL)
-      case "file"  => loadDatasetFromFile(new Path(uri.getPath))
-      case s       => throw new SchemaNotSupportedException(s)
-    }
-  }
-
-  def loadDatasetFromURL(url: java.net.URL): Dataset = {
-    val source = scala.io.Source.fromURL(url)
-    try {
-      yaml.load[Manifest[Dataset]](source.mkString).content
-    } catch {
-      case e: Exception =>
-        logger.error(s"Error while loading data source manifest from URL: $url")
-        throw e
-    } finally {
-      source.close()
-    }
+    datasetRepo.loadResourceFromURI(uri)
   }
 
   def addDataset(ds: Dataset): Unit = {
-    val path = getDatasetDefinitionPath(ds.id)
-
-    if (fileSystem.exists(path))
-      throw new AlreadyExistsException(ds.id)
+    if (datasetRepo.getResourceOpt(ds.id).isDefined)
+      throw new AlreadyExistsException(ds.id.toString, "dataset")
 
     try {
       ds.dependsOn.map(getDataset)
+      ds.remoteSource.map(rs => getVolume(rs.volumeID))
     } catch {
-      case e: DoesNotExistsException =>
-        throw new MissingReferenceException(ds.id, e.datasetID)
+      case e: DoesNotExistException =>
+        throw new MissingReferenceException(
+          ds.id.toString,
+          "dataset",
+          e.id,
+          e.kind
+        )
     }
 
-    saveDataset(ds, path)
-  }
-
-  def saveDataset(ds: Dataset, path: Path): Unit = {
-    val outputStream = fileSystem.create(path)
-
-    try {
-      yaml.save(ds.asManifest, outputStream)
-    } catch {
-      case e: Exception =>
-        outputStream.close()
-        fileSystem.delete(path, false)
-        throw e
-    } finally {
-      outputStream.close()
-    }
+    datasetRepo.addResource(ds)
   }
 
   def deleteDataset(id: DatasetID): Unit = {
@@ -144,12 +113,9 @@ class MetadataRepository(
     if (referencedBy.nonEmpty)
       throw new DanglingReferenceException(referencedBy.map(_.id), id)
 
-    val path = getDatasetDefinitionPath(id)
-    if (!fileSystem.exists(path))
-      throw new DoesNotExistsException(id)
-
     getAllDataPaths(id).foreach(p => fileSystem.delete(p, true))
-    fileSystem.delete(path, false)
+
+    datasetRepo.deleteResource(id)
   }
 
   def purgeDataset(id: DatasetID): Unit = {
@@ -158,17 +124,18 @@ class MetadataRepository(
     // TODO: Purging a dataset that is used by non-empty derivatives should raise an error
   }
 
-  protected def getAllDataPaths(id: DatasetID): Seq[Path] = {
-    val volume = getVolumeFor(id)
-
-    Seq(
-      volume.dataDir.resolve(id.toString),
-      volume.checkpointsDir.resolve(id.toString)
-    )
+  def exportDataset(ds: Dataset, path: Path): Unit = {
+    datasetRepo.saveResourceToFile(ds, path)
   }
 
-  protected def getDatasetDefinitionPath(id: DatasetID): Path = {
-    workspaceLayout.datasetsDir.resolve(id.toString + ".yaml")
+  protected def getAllDataPaths(id: DatasetID): Seq[Path] = {
+    val volumeLayout = getLocalVolume()
+
+    Seq(
+      volumeLayout.dataDir.resolve(id.toString),
+      volumeLayout.cacheDir.resolve(id.toString),
+      volumeLayout.checkpointsDir.resolve(id.toString)
+    )
   }
 
   ////////////////////////////////////////////////////////////////////////////
@@ -188,35 +155,34 @@ class MetadataRepository(
     VolumeLayout(
       datasetsDir = workspaceLayout.localVolumeDir.resolve("datasets"),
       checkpointsDir = workspaceLayout.localVolumeDir.resolve("checkpoints"),
-      dataDir = workspaceLayout.localVolumeDir.resolve("data")
+      dataDir = workspaceLayout.localVolumeDir.resolve("data"),
+      cacheDir = workspaceLayout.localVolumeDir.resolve("cache")
     )
   }
 
-  def getVolumeFor(id: DatasetID): VolumeLayout = {
-    // Only local volumes are supported now
-    getLocalVolume()
+  def loadVolumeFromURI(uri: URI): Volume = {
+    volumeRepo.loadResourceFromURI(uri)
   }
+
+  def getAllVolumeIDs(): Seq[VolumeID] = {
+    volumeRepo.getAllResourceIDs()
+  }
+
+  def getAllVolumes(): Seq[Volume] = {
+    volumeRepo.getAllResources()
+  }
+
+  def getVolume(volumeID: VolumeID): Volume = {
+    volumeRepo.getResource(volumeID)
+  }
+
+  def addVolume(volume: Volume): Unit = {
+    volumeRepo.addResource(volume)
+  }
+
+  def deleteVolume(volumeID: VolumeID): Unit = {
+    // TODO: validate references
+    volumeRepo.deleteResource(volumeID)
+  }
+
 }
-
-/////////////////////////////////////////////////////////////////////////////////////////
-// Exceptions
-/////////////////////////////////////////////////////////////////////////////////////////
-
-class DoesNotExistsException(val datasetID: DatasetID)
-    extends Exception(s"Dataset $datasetID does not exist")
-
-class AlreadyExistsException(val datasetID: DatasetID)
-    extends Exception(s"Dataset $datasetID already exists")
-
-class MissingReferenceException(val fromID: DatasetID, val toID: DatasetID)
-    extends Exception(s"Dataset $fromID refers to non existent dataset $toID")
-
-class SchemaNotSupportedException(val schema: String)
-    extends Exception(s"$schema")
-
-class DanglingReferenceException(
-  val fromIDs: Seq[DatasetID],
-  val toID: DatasetID
-) extends Exception(
-      s"Dataset $toID is referenced by: " + fromIDs.mkString(", ")
-    )