+++ src/main/scala/dev/kamu/cli/commands/AddInteractiveCommand.scala
@@ -37,7 +37,7 @@ class AddInteractiveCommand(
       logger.info("Added dataset")
     } else {
       val path = new Path("./" + dataset.id + ".yaml")
-      metadataRepository.saveDataset(
+      metadataRepository.exportDataset(
         dataset,
         path
       )