+++ src/test/scala/dev/kamu/cli/KamuTestBase.scala
@@ -8,39 +8,21 @@
 
 package dev.kamu.cli
 
-import java.util.UUID
-
 import dev.kamu.core.utils.fs._
 import dev.kamu.core.utils.test.KamuDataFrameSuite
 import org.apache.hadoop.conf.Configuration
-import org.apache.hadoop.fs.{FileSystem, Path}
+import org.apache.hadoop.fs.FileSystem
 import org.scalatest._
 
 trait KamuTestBase extends KamuDataFrameSuite { self: Suite =>
 
   val fileSystem = FileSystem.get(new Configuration())
 
-  def getSystemTempDir(): Path =
-    new Path(System.getProperty("java.io.tmpdir"))
-
-  def getRandomDir(): Path =
-    getSystemTempDir()
-      .resolve("kamu-test-" + UUID.randomUUID.toString)
-
   def withEmptyDir[T](func: KamuTestAdapter => T): T = {
-    val testDir = getRandomDir()
-    fileSystem.mkdirs(testDir)
-
-    try {
-      val config = KamuConfig(
-        workspaceRoot = testDir
-      )
-
+    Temp.withRandomTempDir(fileSystem, "kamu-test-") { tempDir =>
+      val config = KamuConfig(workspaceRoot = tempDir)
       val kamu = new KamuTestAdapter(config, fileSystem, spark)
-
       func(kamu)
-    } finally {
-      fileSystem.delete(testDir, true)
     }
   }