+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.0] - 2020-01-12
+### Added
+- Experimental support for remote S3 volumes
+
+## [0.7.1] - 2019-12-29
+### Changed
+- Bumped ingest version
+
+## [0.7.0] - 2019-12-29
+### Changed
+- Using snake_case dataset vocabulary
+
 ## [0.6.0] - 2019-12-15
 ### Added
 - Richer set of CSV reader options